liburdma: Do not send DPDK log messages to syslog
If urdmad running under systemd, stderr is automatically logged to the
systemd journal and thus also logging to syslog (which DPDK does by
default) results in duplicate log messages. By preserving stderr, we
ensure that any error messages printed by a library (such as glibc heap
corruption dumps) make it to the journal.

For the case of liburdma, it is distracting to have these messages sent
to syslog. Let these be printed in stderr.

Signed-off-by: Patrick MacArthur <patrick@patrickmacarthur.net>