[WIP] src/kmod: Use dev_{dbg,info,notice,warn,err}() instead of pr_*()
From Jason Gunthorpe:
> It is not appropriate for a driver to use pr_*....

Also we get a prefix with the device name for free.

Take this opportunity to change the severity level of some of our
messages, and remove some of our other custom debug logic (the DBG_*
macros along with hand-printing IPv4 addresses which can be done with
the "%pI4" format code).